+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `CTAButtonV2`
 - Add `design-system` page
 - Add `aereo` rewrite to Mauricio's Fellowship project in now.json
+- Add `HeaderNav` component
 
 ## [1.8.1] - 2020-04-01